% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idcnrba-package.R
\docType{package}
\name{idcnrba-package}
\alias{idcnrba}
\alias{idcnrba-package}
\title{idcnrba: Interactive Application for Analyzing Representativeness and Nonresponse Bias}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides access to the Idea Data Center (IDC) application for conducting nonresponse bias analysis (NRBA). The IDC NRBA app is an interactive, browser-based Shiny application that can be used to analyze survey data with respect to response rates, representativeness, and nonresponse bias. This app provides a user-friendly interface to statistical methods implemented by the 'nrba' package. Krenzke, Van de Kerckhove, and Mohadjer (2005) \url{http://www.asasrms.org/Proceedings/y2005/files/JSM2005-000572.pdf} and Lohr and Riddles (2016) \url{https://www150.statcan.gc.ca/n1/en/pub/12-001-x/2016002/article/14677-eng.pdf?st=q7PyNsGR} provide an overview of the statistical methods implemented in the application.
}
\author{
\strong{Maintainer}: Ben Schneider \email{BenjaminSchneider@westat.com} (\href{https://orcid.org/0000-0002-0406-8470}{ORCID})

Authors:
\itemize{
  \item Tamara Nimkoff \email{TamaraNimkoff@westat.com}
  \item Anthony Fucci \email{AnthonyFucci@westat.com}
  \item Andy Cruse \email{AndyCruse@westat.com}
  \item Alexander Cates
  \item Jim Green \email{JimGreen@westat.com}
}

Other contributors:
\itemize{
  \item Westat [copyright holder]
}

}
\keyword{internal}
